refactor: remove unused ABIs
