--- lib/abi/public-resolver.json
@@ -1,25 +1,0 @@
-[
-  {
-    "constant": false,
-    "inputs": [
-      { "internalType": "bytes32", "name": "node", "type": "bytes32" },
-      { "internalType": "uint256", "name": "coinType", "type": "uint256" },
-      { "internalType": "bytes", "name": "a", "type": "bytes" }
-    ],
-    "name": "setAddr",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      { "internalType": "bytes32", "name": "node", "type": "bytes32" },
-      { "internalType": "uint256", "name": "coinType", "type": "uint256" }
-    ],
-    "name": "addr",
-    "outputs": [{ "internalType": "bytes", "name": "", "type": "bytes" }],
-    "stateMutability": "view",
-    "type": "function"
-  }
-]
