+++ lib/utils/blockchain-txs.ts
@@ -32,7 +32,7 @@ import {
 } from "../wallet/txError";
 import { getNameRegistrationSecret } from "@/lib/name-registration/localStorage";
 import { parseAccount } from "viem/utils";
-import DomainResolverABI from "../abi/resolver.json";
+import DomainResolverABI from "../abi/offchain-resolver.json";
 import { normalize } from "viem/ens";
 import { cryptocurrencies } from "../domain-page";
 import { getCoderByCoinName } from "@ensdomains/address-encoder";
